% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{fahrenheit_to_celsius}
\alias{fahrenheit_to_celsius}
\title{fahrenheit to celsius}
\usage{
fahrenheit_to_celsius(temp_F)
}
\value{
celsius transformed according to fahernheit
}
\description{
This function is used to transfer fahernheit to celsius
}
