% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_healthcare.R
\docType{data}
\name{german_healthcare}
\alias{german_healthcare}
\title{Autosales data set.}
\format{
An object of class \code{data.frame} with 27327 rows and 43 columns.
}
\usage{
german_healthcare
}
\description{
This is a large data set. There are altogether 27,326 observations.
The number of observations ranges from 1 to 7.
(Frequencies are: 1=1525, 2=2158, 3=825, 4=926, 5=1051, 6=1000, 7=987).
Note, the variable NUMOBS below tells how many observations there are for each person.
Thisvariable is repeated in each row of the data for the person.
}
\keyword{datasets}
